refact: 캐싱 기능을 담당하는 객체를 별도의 파일로 추출
