+++ src/apis/apiClient.js
@@ -1,44 +1,5 @@
 import axios from 'axios';
-// TODO: extract to a separate file
-class Cache {
-  #dataMap = {};
-  #expireTimeInSec;
-
-  constructor(expireTimeInSec) {
-    this.#expireTimeInSec = expireTimeInSec || 0;
-  }
-
-  get(key) {
-    return this.#dataMap[key];
-  }
-
-  set(key, data, expireTimeInSec) {
-    const cacheData = new Data(data, expireTimeInSec || this.#expireTimeInSec);
-    this.#dataMap[key] = cacheData;
-  }
-}
-
-class Data {
-  #data;
-  #expireTimeInSec;
-  #createdAt;
-
-  constructor(data, expireTimeInSec) {
-    this.#data = data;
-    this.#expireTimeInSec = expireTimeInSec;
-    this.#createdAt = new Date();
-  }
-
-  get data() {
-    return this.#data;
-  }
-
-  get isExpired() {
-    const now = new Date();
-    const differenceSec = (now.getTime() - this.#createdAt.getTime()) / 1000;
-    return differenceSec >= this.#expireTimeInSec;
-  }
-}
+import { Cache } from './cache';
 
 class ApiClient {
   #options = {};
